SingleStrategiesFinal
Finally finished the single strategies. Everything seems to be working smoothly as tested with prints. I am still not completely satisfied with the code modularity and tidyness:
E.g: The q_learning agent strategy is split between the agent class, the experiment class and the main function. It should be possible to just put them in a single function running as called by the run single experiment.
Also Periodic function starts with just two Waits instead of three which scares me to death that it might be a symptom of something bigger. Then again the modulus method and the individual add trial in this strategy is a bit unorthodox so it might have to do with that. I am also not sure if I should create a strategy class and run the strategies from there. Also, each strategy should be called with their own arguments and right now it is a bit of a mess. Regardless. It is working, both on the surface and below and I do understand the logic! I am proud!